Fix issue with using 40-bit encryption and PDF file creation.

+++ source/cpp/ctlpdf.cpp
@@ -59,13 +59,16 @@ DTWAIN_BOOL DLLENTRY_DEF DTWAIN_SetPDFProducer(DTWAIN_SOURCE Source, LPCTSTR lpP
 
 DTWAIN_BOOL DLLENTRY_DEF DTWAIN_SetPDFEncryption(DTWAIN_SOURCE Source, DTWAIN_BOOL bUseEncryption,
                                                  LPCTSTR lpszUser, LPCTSTR lpszOwner,
-                                                 LONG Permissions, DTWAIN_BOOL UseStrongEncryption)
+                                                 DWORD Permissions, DTWAIN_BOOL UseStrongEncryption)
 {
     LOG_FUNC_ENTRY_PARAMS((Source, bUseEncryption, lpszUser, lpszOwner, Permissions, UseStrongEncryption))
     auto [pHandle, pSource] = VerifyHandles(Source);
     const CTL_StringType owner = lpszOwner?lpszOwner:_T("");
     const CTL_StringType user = lpszUser?lpszUser:_T("");
 
+    // Even though the Permissions parameter is an unsigned 32-bit value from the user, 
+    // this will be "converted" to a 32-bit signed integer internally, which is what the PDF 
+    // specification requires.
     pSource->SetPDFEncryption(bUseEncryption?true:false, user, owner, Permissions, UseStrongEncryption?true:false);
     LOG_FUNC_EXIT_NONAME_PARAMS(true)
     CATCH_BLOCK_LOG_PARAMS(false)
